+++ metrics_sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History: opentelemetry-metrics-sdk
 
+### v0.5.0 / 2025-01-08
+
+* ADDED: Add synchronous gauge
+
+### v0.4.1 / 2024-12-04
+
+* FIXED: Handle float value in NumberDataPoint
+
+### v0.4.0 / 2024-11-20
+
+* ADDED: Update metrics configuration patch
+
 ### v0.3.0 / 2024-10-22
 
 * ADDED: Add basic metrics view